refactor: candidature manager
