+++ app/views/manager/candidatures/_candidature.html.erb
@@ -15,9 +15,9 @@
 
   <% if action_name != "show" %>
   <div class="flex flex-row">
-      <%= link_to t('views.manager.candidatures.edit'), edit_candidature_path(candidature),
+      <%= link_to t('views.manager.candidatures.edit'), edit_manager_candidature_path(candidature),
                   class:"mr-2 items-center justify-center focus:outline-none text-white bg-yellow-400 hover:bg-yellow-500 focus:ring-4 focus:ring-yellow-300 font-medium rounded-lg text-sm px-5 py-2.5 me-2 mb-2 dark:focus:ring-yellow-900" %>
-      <%= link_to  candidature_path(candidature), data: { turbo_method: :delete, turbo_confirm: "Are you sure?" },
+      <%= link_to  manager_candidature_path(candidature), data: { turbo_method: :delete, turbo_confirm: "Are you sure?" },
                    class:"mr-2 items-center justify-center focus:outline-none text-white bg-red-700 hover:bg-red-800 focus:ring-4 focus:ring-red-300 font-medium rounded-lg text-sm px-5 py-2.5 me-2 mb-2 dark:bg-red-600 dark:hover:bg-red-700 dark:focus:ring-red-900" do %>
        <%= t('views.manager.candidatures.delete') %>
       <% end %>
