+++ app/views/manager/candidatures/_candidature.json.jbuilder
