+++ app/views/manager/candidatures/_form.html.erb
@@ -1,5 +1,5 @@
 <div>
-  <%= simple_form_for @candidature do |f| %>
+  <%= simple_form_for [:manager, @candidature] do |f| %>
     <div>
       <%= f.input :company_name %>
     </div>
@@ -10,7 +10,7 @@
     <div class="flex items-center">
       <%= f.button :submit, class: "cursor-pointer" %>
       <div>
-        <%= link_to t('views.manager.candidatures.back'), candidatures_path,
+        <%= link_to t('views.manager.candidatures.back'), manager_candidatures_path,
                     class: "mr-2 items-center justify-center focus:outline-none text-white bg-green-700 hover:bg-green-800 focus:ring-4 focus:ring-green-300 font-medium rounded-lg text-sm px-5 py-2.5 me-2 mb-2 dark:bg-green-600 dark:hover:bg-green-700 dark:focus:ring-green-800" %>
       </div>
     </div>
