+++ app/views/manager/candidatures/edit.html.erb
