+++ app/views/manager/candidatures/index.html.erb
@@ -5,7 +5,7 @@
 
   <div>
     <div class="py-4">
-      <%= link_to t('views.manager.candidatures.index.add_candidature'), new_candidature_path,
+      <%= link_to t('views.manager.candidatures.index.add_candidature'), new_manager_candidature_path,
                   class:"text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 me-2 mb-2 dark:bg-blue-600 dark:hover:bg-blue-700 focus:outline-none dark:focus:ring-blue-800" %>
 
     </div>
