+++ app/views/manager/candidatures/index.json.jbuilder
