+++ app/views/manager/candidatures/new.html.erb
