+++ app/views/manager/candidatures/show.json.jbuilder
